moved the location of BungieComponents into the type folder
