+++ src/util/get-light-level.ts
@@ -1,9 +1,8 @@
 // this function gets the light level of a particular item
 
-import {
-	BungieComponents,
-	requestItemInstanceComponent,
-} from "./bungie-api-request";
+import { requestItemInstanceComponent } from "./bungie-api-request";
+
+import { BungieComponents } from "@/types/bungie-components";
 
 export async function getLightLevel(itemInstanceId: string): Promise<number> {
 	const data = await requestItemInstanceComponent(
