+++ src/util/get-power-data.ts
@@ -9,10 +9,7 @@ import { levelCaps } from "@/types/power-level-caps";
 
 // this functions gets the data necessary to create the power page
 
-export async function fetchPowerData(
-	membershipType: number,
-	membershipId: string,
-): Promise<PowerPageProps> {
+export async function fetchPowerData(): Promise<PowerPageProps> {
 	// this has every item on your account
 	const combinedData = await getAllItems();
 
@@ -74,7 +71,7 @@ export async function fetchPowerData(
 	);
 
 	// gets your account's light level bonus
-	const lightLevelBonus = await getPowerBonus(membershipType, membershipId);
+	const lightLevelBonus = await getPowerBonus();
 
 	// calculates the base light level using the highest light items
 	const lightLevel =
